# Optimize wfn.HasWildcard a bit.