@skogtwin 2020-04-07
Yeah, build failure is unrelated to the PR, `golangci-lint run` complains about imports for some reason; we need to fix .travis.yaml, I guess (https://travis-ci.com/github/facebookincubator/nvdtools/jobs/316473574)